split contextualisation

+++ src/neural_bandits/bandit_algorithms/abstract_bandit.py
@@ -11,18 +11,18 @@ class AbstractBandit(ABC):
         self.n_features = n_features
 
     @abstractmethod
-    def predict(self, feature_vectors: List[npt.NDArray[np.int_]]) -> List[int]:
+    def predict(self, feature_vectors: List[npt.NDArray[np.float64]]) -> List[int]:
         """Predict a list of multiple sets of contextualised actions"""
         pass
 
     @abstractmethod
-    def select_arm(self, contextualised_actions: npt.NDArray[np.int_]) -> int:
+    def select_arm(self, contextualised_actions: npt.NDArray[np.float64]) -> int:
         """Select an arm based on the contextualised actions"""
         pass
 
     @abstractmethod
     def update_step(
-        self, reward: float, chosen_contextualised_action: npt.NDArray[np.int_]
+        self, reward: float, chosen_contextualised_action: npt.NDArray[np.float64]
     ) -> None:
         """Perform a single update step"""
         pass
@@ -31,7 +31,7 @@ class AbstractBandit(ABC):
     def update_batch(
         self,
         rewards: List[float],
-        chosen_contextualised_actions: List[npt.NDArray[np.int_]],
+        chosen_contextualised_actions: List[npt.NDArray[np.float64]],
     ) -> None:
         """Perform a batch update"""
         pass
